requirements.txt
notion-client
watchdog
markdown2
python-dotenv
beautifulsoup4
